Added authorization to product deletion route.

+++ controllers/api/productRoutes.js
@@ -9,7 +9,7 @@ const withAuth = require('../../utils/auth');
 
 // Add a new product if a user is logged in using withAuth middleware
 router.post('/', withAuth, async (req, res) => {
-    
+
     // Req.body should look like:
     // {
     //      "product_name": "something",
@@ -22,7 +22,7 @@ router.post('/', withAuth, async (req, res) => {
     try {
 
         // Get category data by category name from database
-        const category = await Category.findOne({ 
+        const category = await Category.findOne({
             where: {
                 category_name: req.body.category_name
             }
@@ -38,7 +38,7 @@ router.post('/', withAuth, async (req, res) => {
         const product = await Product.create(req.body);
 
         res.status(200).json(product);
-        
+
     } catch (error) {
         res.status(500).json(error);
     }
@@ -53,7 +53,7 @@ router.get('/', async (req, res) => {
         // res.status(200).json(productData);
         const allProducts = productData.map((product) => product.get({ plain: true }));
         console.log(productData);
-        res.render('product', {allProducts});
+        res.render('product', { allProducts });
     } catch (err) {
         res.status(500).json(err);
     }
@@ -68,7 +68,7 @@ router.get('/', async (req, res) => {
 //             },
 //             include: [ { model: Category }, { model: Product } ],
 //         });
-        
+
 //         if(!productData) {
 //             res.status(404).json({ message: 'No product found with that ID'});
 //             return;
@@ -80,16 +80,28 @@ router.get('/', async (req, res) => {
 //     }
 // });
 
-router.delete('/:id', async (req, res) => {
+router.delete('/:id', withAuth, async (req, res) => {
 
     try {
+
+        // Get product-being-deleted's owner ID
+        const productToDelete = await Product.findByPk(req.params.id);
+        const productOwnerID = productToDelete.dataValues.user_id;
+
+        // Check to make sure the person deleting the product is the owner of the product
+        if (productOwnerID !== req.session.user_id) {
+            res.status(403).json({ message: "You cannot delete products that do not belong to you" });
+            return;
+        }
+
+        // Attempt to delete the product
         const deletedProduct = await Product.destroy({
             where: { 
                 id: req.params.id
             }
         });
 
-        res.status(200).json({ message: 'Product deleted successfully'});
+        res.status(200).json({ message: 'Product deleted successfully' });
 
     } catch (error) {
         res.status(500).json(error);
